Code/requirements.txt
astroid==2.5.6
autopep8==1.5.6
bitarray==2.0.1
cycler==0.10.0
Cython==0.29.23
isort==5.8.0
jpegio==0.2.2
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
matplotlib==3.4.2
mccabe==0.6.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pycodestyle==2.7.0
pylint==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
reedsolo==1.5.4
scipy==1.6.3
six==1.16.0
toml==0.10.2
wrapt==1.12.1
